refactor: extract @kolu/solid-pierre package
Move the three Pierre wrappers (PierreFileTree, PierreDiffView,
PierreFileView) out of packages/client/src/ui/ and into a new
packages/solid-pierre/ workspace package. Encapsulating the imperative
mount/render/cleanUp dance behind one Solid-native interface means a
future Pierre-backed surface (commit timeline, search results,
preview pane) reuses the safe wrapper instead of repeating the same
scaffolding — and the same bugs.

Closes #807. Wraps the issue's "What lives inside" scope:

- Lifecycle: onMount + onCleanup own Pierre's instance ref.
- Reactivity bridge: createEffect(on(...)) setters with correct
  dependency tracking. Initial searchQuery is applied at mount so a
  pre-mount value isn't silently dropped.
- Error boundary: every imperative .render() and parse call is
  try/caught and routed to a required onError prop. Pierre throws no
  longer escape Solid's <ErrorBoundary> (which only catches errors
  during *Solid* render) and crash the parent tree.
- Type-safe context menus: Pierre's composition.contextMenu hook is
  exposed as a typed FileTreeContextMenu prop, not a stringly-keyed
  escape hatch.

Kolu-specific bits stay in packages/client/:
- pierreTheme.ts (kolu palette CSS overrides + nix/hs sprite sheet).
- useLineSelection + CodeContextMenu (kolu's "Copy <path>:<line>" UX
  is a product feature, not a Pierre wrapper concern).
- toGitStatusEntries + renderTreeContextMenu live in the new
  pierreAdapters.ts since they bridge kolu types and the Pierre API.

Virtualization variants (VirtualizedFile, VirtualizedFileDiff) are
out of scope — no kolu consumer needs them yet; they'll land with the
#514 Phase 8 work that motivated them.
