+++ packages/client/package.json
@@ -15,6 +15,7 @@
     "@corvu/drawer": "^0.2.2",
     "@corvu/resizable": "^0.2.5",
     "@corvu/tooltip": "^0.2.2",
+    "@kolu/solid-pierre": "workspace:*",
     "@kolu/surface": "workspace:*",
     "@orpc/client": "^1.13.13",
     "@pierre/diffs": "^1.1.20",
