+++ packages/solid-pierre/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@kolu/solid-pierre",
+  "version": "0.1.0",
+  "private": true,
+  "license": "AGPL-3.0-or-later",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@types/node": "^22.0.0",
+    "typescript": "^5.8.0"
+  },
+  "dependencies": {
+    "@pierre/diffs": "^1.1.20",
+    "@pierre/trees": "^1.0.0-beta.3",
+    "solid-js": "^1.9.0"
+  }
+}
